refactor(slurm): rename SlurmSSHAdapter→SlurmSSHClient + split into clients/ (#193)
Closes the last item in #193's "split oversized modules" cleanup.

## File move + class rename

* ``src/srunx/slurm/ssh.py`` (1887 LOC) → ``src/srunx/slurm/clients/ssh.py``,
  matching the ``clients/local.py`` sibling landed in #194.
* ``SlurmSSHAdapter`` → ``SlurmSSHClient`` (and ``SlurmSSHAdapterSpec`` →
  ``SlurmSSHClientSpec``) so the class name reads as a peer of
  ``LocalClient`` rather than spelling "Adapter" against a sibling that
  spells "Client". Disambiguates against
  ``srunx.ssh.core.client.SSHSlurmClient`` per the issue's naming-symmetry
  note (the ``Slurm`` prefix order keeps the two distinct).

## Internal helper extraction

The single 1887-LOC class is now ~1100 LOC after pulling out four
cohesive helper modules under ``slurm/clients/``:

  src/srunx/slurm/clients/
  ├── ssh.py             SlurmSSHClient class — owns connection lifecycle,
  │                      submit/run lifecycle, _io_lock discipline.
  ├── _ssh_helpers.py    Shared free helpers (_run_slurm_cmd,
  │                      _validate_identifier, _MountsOnly, regex consts).
  ├── _ssh_types.py      SlurmSSHClientSpec, SSHMonitorTimeoutError,
  │                      _resolve_monitor_timeout_default.
  ├── _ssh_resources.py  sinfo+squeue resource queries (~200 LOC).
  ├── _ssh_queries.py    squeue+sacct+scontrol job listing/parsing (~290 LOC).
  └── _ssh_recording.py  best-effort jobs/job_state_transitions DB writes.

Each helper module takes the client as its first argument — class
methods (``get_resources`` / ``list_jobs`` / ``queue_by_ids`` / etc.)
are 1-line forwards. The parsing and resource logic is now
independently reviewable / testable; the connection lock + lifecycle
stay on the class where they belong.

## Caller updates (38 files)

  37 ``from srunx.slurm.ssh import …`` → ``from srunx.slurm.clients.ssh
  import …`` updates across src + tests, and every type annotation /
  isinstance check / patch path swapped to the new ``SlurmSSHClient`` /
  ``SlurmSSHClientSpec`` names. Internal helpers
  (``_validate_identifier`` / ``_run_slurm_cmd``) moved to
  ``slurm.clients._ssh_helpers``; ``slurm/accounting.py`` and
  ``slurm/partitions.py`` updated their lazy imports accordingly.

  Test ``@patch`` paths follow the actual call sites:

    srunx.slurm.clients.ssh._run_slurm_cmd
      → srunx.slurm.clients._ssh_queries._run_slurm_cmd        (list_jobs/get_job/queue_by_ids tests)
      → srunx.slurm.clients._ssh_resources._run_slurm_cmd      (resource tests)
    srunx.slurm.clients.ssh._validate_identifier
      → srunx.slurm.clients._ssh_resources._validate_identifier (resource tests)

## Side cleanups

* ``_ssh_recording.py`` ditched two ``# type: ignore[arg-type]`` by typing
  ``transport_type`` / ``submission_source`` as the actual ``TransportType``
  / ``SubmissionSource`` ``Literal`` aliases.
* ``run()``'s ``source_path = None  # type: ignore[assignment]`` replaced
  with a proper ``Path | None`` annotation on the local.

The remaining 2 ``# type: ignore[arg-type]`` on ``_record_job_submission``
preserve the boundary cast for ``self.submission_source`` (plain ``str``
because callers pass ``"mcp"`` which is not in the
``SubmissionSource = Literal["cli", "web", "workflow"]`` Literal).
Documented inline.

The ``# noqa: BLE001`` catches that survive are all justified-and-documented
defensive catches at signal-handler / connection-probe / callback /
top-level-handler boundaries (5 sites total, each with an explanatory
comment).

## Verification

* 2043 tests pass (no regressions).
* ruff + mypy clean (188 source files).
* All 38 callers + tests updated mechanically; smoke-tested every
  affected import chain.

Refs #193 (final split — issue can be closed when this lands).
