+++ src/srunx/slurm/clients/__init__.py
@@ -5,8 +5,8 @@ Each module here is a transport-specific implementation of the
 
 - :mod:`srunx.slurm.clients.local` — :class:`LocalClient`, the
   in-process ``sbatch`` / ``squeue`` / ``scancel`` driver.
-- :mod:`srunx.slurm.clients.ssh` (planned) — the SSH-transport
-  sibling, currently still in :mod:`srunx.slurm.ssh`.
+- :mod:`srunx.slurm.clients.ssh` — :class:`SlurmSSHClient`, the
+  SSH-transport sibling.
 
 Public consumers should generally import :class:`Slurm` from
 :mod:`srunx.slurm.local` (the legacy wrapper that wires DB recording +
