+++ src/srunx/web/routers/resources.py
@@ -7,7 +7,7 @@ from typing import Any
 import anyio
 from fastapi import APIRouter, Depends, HTTPException
 
-from srunx.slurm.ssh import SlurmSSHAdapter
+from srunx.slurm.clients.ssh import SlurmSSHClient
 
 from ..deps import get_adapter
 
@@ -17,7 +17,7 @@ router = APIRouter(prefix="/api/resources", tags=["resources"])
 @router.get("")
 async def get_resources(
     partition: str | None = None,
-    adapter: SlurmSSHAdapter = Depends(get_adapter),
+    adapter: SlurmSSHClient = Depends(get_adapter),
 ) -> list[dict[str, Any]]:
     """Get current resource availability via SSH."""
     try:
